updated database, and completed server
